mp_integer: move some tests from 02 to 03, as the latest travis build timed out.
