+++ tests/mp_integer_02.cpp
@@ -51,7 +51,6 @@ see https://www.gnu.org/licenses/. */
 #include <unordered_map>
 #include <vector>
 
-#include "../src/binomial.hpp"
 #include "../src/config.hpp"
 #include "../src/debug_access.hpp"
 #include "../src/environment.hpp"
@@ -2127,290 +2126,6 @@ struct hash_tester
 BOOST_AUTO_TEST_CASE(mp_integer_hash_test)
 {
 	boost::mpl::for_each<size_types>(hash_tester());
-}
-
-struct next_prime_tester
-{
-	template <typename T>
-	void operator()(const T &)
-	{
-		typedef mp_integer<T::value> int_type;
-		int_type n;
-		BOOST_CHECK_EQUAL(n.nextprime(),2);
-		n = 2;
-		BOOST_CHECK_EQUAL(n.nextprime(),3);
-		n = 3;
-		BOOST_CHECK_EQUAL(n.nextprime(),5);
-		n = 7901;
-		BOOST_CHECK_EQUAL(n.nextprime(),7907);
-		n = -1;
-		BOOST_CHECK_THROW(n.nextprime(),std::invalid_argument);
-		// Random tests.
-		std::uniform_int_distribution<int> ud(std::numeric_limits<int>::lowest(),std::numeric_limits<int>::max());
-		std::uniform_int_distribution<int> promote_dist(0,1);
-		mpz_raii m;
-		for (int i = 0; i < ntries; ++i) {
-			auto tmp = ud(rng);
-			n = tmp;
-			if (promote_dist(rng) && n.is_static()) {
-				n.promote();
-			}
-			if (tmp < 0) {
-				BOOST_CHECK_THROW(n.nextprime(),std::invalid_argument);
-				continue;
-			}
-			::mpz_set_si(&m.m_mpz,static_cast<long>(tmp));
-			::mpz_nextprime(&m.m_mpz,&m.m_mpz);
-			BOOST_CHECK_EQUAL(boost::lexical_cast<std::string>(n.nextprime()),mpz_lexcast(m));
-		}
-	}
-};
-
-BOOST_AUTO_TEST_CASE(mp_integer_next_prime_test)
-{
-	boost::mpl::for_each<size_types>(next_prime_tester());
-}
-
-struct probab_prime_p_tester
-{
-	template <typename T>
-	void operator()(const T &)
-	{
-		typedef mp_integer<T::value> int_type;
-		int_type n;
-		BOOST_CHECK(n.probab_prime_p() == 0);
-		n = 1;
-		BOOST_CHECK(n.probab_prime_p() == 0);
-		n = 2;
-		BOOST_CHECK(n.probab_prime_p() != 0);
-		n = 3;
-		BOOST_CHECK(n.probab_prime_p() != 0);
-		n = 5;
-		BOOST_CHECK(n.probab_prime_p() != 0);
-		n = 11;
-		BOOST_CHECK(n.probab_prime_p() != 0);
-		n = 16;
-		BOOST_CHECK(n.probab_prime_p() != 2);
-		n = 7901;
-		BOOST_CHECK(n.probab_prime_p() != 0);
-		n = 7907;
-		BOOST_CHECK(n.probab_prime_p(5) != 0);
-		n = -1;
-		BOOST_CHECK_THROW(n.probab_prime_p(),std::invalid_argument);
-		n = 5;
-		BOOST_CHECK_THROW(n.probab_prime_p(0),std::invalid_argument);
-		BOOST_CHECK_THROW(n.probab_prime_p(-1),std::invalid_argument);
-	}
-};
-
-BOOST_AUTO_TEST_CASE(mp_integer_probab_prime_p_test)
-{
-	boost::mpl::for_each<size_types>(probab_prime_p_tester());
-}
-
-struct integer_sqrt_tester
-{
-	template <typename T>
-	void operator()(const T &)
-	{
-		typedef mp_integer<T::value> int_type;
-		int_type n;
-		BOOST_CHECK(n.sqrt() == 0);
-		n = 1;
-		BOOST_CHECK(n.sqrt() == 1);
-		n = 2;
-		BOOST_CHECK(n.sqrt() == 1);
-		n = 3;
-		BOOST_CHECK(n.sqrt() == 1);
-		n = 4;
-		BOOST_CHECK(n.sqrt() == 2);
-		n = 5;
-		BOOST_CHECK(n.sqrt() == 2);
-		// Random tests.
-		std::uniform_int_distribution<int> ud(std::numeric_limits<int>::lowest(),std::numeric_limits<int>::max());
-		std::uniform_int_distribution<int> promote_dist(0,1);
-		mpz_raii m;
-		for (int i = 0; i < ntries; ++i) {
-			auto tmp = ud(rng);
-			n = tmp;
-			if (promote_dist(rng) && n.is_static()) {
-				n.promote();
-			}
-			if (tmp < 0) {
-				BOOST_CHECK_THROW(n.sqrt(),std::invalid_argument);
-				continue;
-			}
-			::mpz_set_si(&m.m_mpz,static_cast<long>(tmp));
-			::mpz_sqrt(&m.m_mpz,&m.m_mpz);
-			BOOST_CHECK_EQUAL(boost::lexical_cast<std::string>(n.sqrt()),mpz_lexcast(m));
-		}
-	}
-};
-
-BOOST_AUTO_TEST_CASE(mp_integer_integer_sqrt_test)
-{
-	boost::mpl::for_each<size_types>(integer_sqrt_tester());
-}
-
-struct factorial_tester
-{
-	template <typename T>
-	void operator()(const T &)
-	{
-		typedef mp_integer<T::value> int_type;
-		int_type n;
-		BOOST_CHECK(n.factorial() == 1);
-		n = 1;
-		BOOST_CHECK(n.factorial() == 1);
-		n = 2;
-		BOOST_CHECK(n.factorial() == 2);
-		n = 3;
-		BOOST_CHECK(n.factorial() == 6);
-		n = 4;
-		BOOST_CHECK(n.factorial() == 24);
-		n = 5;
-		BOOST_CHECK(n.factorial() == 24 * 5);
-		// Random tests.
-		std::uniform_int_distribution<int> ud(-1000,1000);
-		std::uniform_int_distribution<int> promote_dist(0,1);
-		mpz_raii m;
-		for (int i = 0; i < ntries; ++i) {
-			auto tmp = ud(rng);
-			n = tmp;
-			if (promote_dist(rng) && n.is_static()) {
-				n.promote();
-			}
-			if (tmp < 0) {
-				BOOST_CHECK_THROW(n.factorial(),std::invalid_argument);
-				continue;
-			}
-			::mpz_set_si(&m.m_mpz,static_cast<long>(tmp));
-			::mpz_fac_ui(&m.m_mpz,static_cast<unsigned long>(tmp));
-			BOOST_CHECK_EQUAL(boost::lexical_cast<std::string>(n.factorial()),mpz_lexcast(m));
-			BOOST_CHECK_EQUAL(boost::lexical_cast<std::string>(math::factorial(n)),mpz_lexcast(m));
-		}
-	}
-};
-
-BOOST_AUTO_TEST_CASE(mp_integer_factorial_test)
-{
-	boost::mpl::for_each<size_types>(factorial_tester());
-}
-
-struct binomial_tester
-{
-	template <typename T>
-	void operator()(const T &)
-	{
-		typedef mp_integer<T::value> int_type;
-		BOOST_CHECK((has_binomial<int_type,int_type>::value));
-		BOOST_CHECK((has_binomial<int_type,int>::value));
-		BOOST_CHECK((has_binomial<int_type,unsigned>::value));
-		BOOST_CHECK((has_binomial<int_type,long>::value));
-		BOOST_CHECK((has_binomial<int_type,char>::value));
-		int_type n;
-		BOOST_CHECK(n.binomial(0) == 1);
-		BOOST_CHECK(n.binomial(1) == 0);
-		n = 1;
-		BOOST_CHECK(n.binomial(1) == 1);
-		n = 5;
-		BOOST_CHECK(n.binomial(3) == 10);
-		n = -5;
-		BOOST_CHECK(n.binomial(int_type(4)) == 70);
-		BOOST_CHECK((has_binomial<int_type,int>::value));
-		BOOST_CHECK((has_binomial<int,int_type>::value));
-		BOOST_CHECK((std::is_same<int_type,decltype(math::binomial(int_type{},0))>::value));
-		BOOST_CHECK((std::is_same<decltype(math::binomial(int_type{},0)),int_type>::value));
-		BOOST_CHECK((has_binomial<int_type,double>::value));
-		BOOST_CHECK((has_binomial<double,int_type>::value));
-		BOOST_CHECK((std::is_same<double,decltype(math::binomial(int_type{},0.))>::value));
-		BOOST_CHECK((std::is_same<decltype(math::binomial(int_type{},0.)),double>::value));
-		BOOST_CHECK((has_binomial<int_type,int_type>::value));
-		BOOST_CHECK((std::is_same<int_type,decltype(math::binomial(int_type{},int_type{}))>::value));
-		// Random tests.
-		std::uniform_int_distribution<int> ud(-1000,1000);
-		std::uniform_int_distribution<int> promote_dist(0,1);
-		mpz_raii m;
-		for (int i = 0; i < ntries; ++i) {
-			auto tmp1 = ud(rng), tmp2 = ud(rng);
-			n = tmp1;
-			if (promote_dist(rng) && n.is_static()) {
-				n.promote();
-			}
-			if (tmp2 < 0) {
-				// NOTE: we cannot check this case with GMP, defer to some tests below.
-				BOOST_CHECK_NO_THROW(n.binomial(tmp2));
-				continue;
-			}
-			::mpz_set_si(&m.m_mpz,static_cast<long>(tmp1));
-			::mpz_bin_ui(&m.m_mpz,&m.m_mpz,static_cast<unsigned long>(tmp2));
-			BOOST_CHECK_EQUAL(boost::lexical_cast<std::string>(n.binomial(tmp2)),mpz_lexcast(m));
-			// int -- integral.
-			BOOST_CHECK_EQUAL(math::binomial(n,tmp2),n.binomial(tmp2));
-			// integral -- int.
-			BOOST_CHECK_EQUAL(math::binomial(tmp2,n),int_type(tmp2).binomial(n));
-			// integral -- integral.
-			BOOST_CHECK_EQUAL(math::binomial(tmp2,tmp1),integer(tmp2).binomial(tmp1));
-			// int -- double.
-			BOOST_CHECK_EQUAL(math::binomial(n,static_cast<double>(tmp2)),math::binomial(double(n),static_cast<double>(tmp2)));
-			// double -- int.
-			BOOST_CHECK_EQUAL(math::binomial(static_cast<double>(tmp2),n),math::binomial(static_cast<double>(tmp2),double(n)));
-			BOOST_CHECK_EQUAL(n.binomial(tmp2),n.binomial(int_type(tmp2)));
-			BOOST_CHECK_EQUAL(n.binomial(long(tmp2)),n.binomial(int_type(tmp2)));
-			BOOST_CHECK_EQUAL(n.binomial((long long)(tmp2)),n.binomial(int_type(tmp2)));
-			BOOST_CHECK_EQUAL(n.binomial((unsigned long)(tmp2)),n.binomial(int_type(tmp2)));
-			BOOST_CHECK_EQUAL(n.binomial((unsigned long long)(tmp2)),n.binomial(int_type(tmp2)));
-		}
-		BOOST_CHECK_THROW(n.binomial(std::numeric_limits<unsigned long>::max() + int_type(1)),std::invalid_argument);
-		// Negative k.
-		BOOST_CHECK_EQUAL(int_type{-3}.binomial(-4),-3);
-		BOOST_CHECK_EQUAL(int_type{-3}.binomial(-10),-36);
-		BOOST_CHECK_EQUAL(int_type{-3}.binomial(-1),0);
-		BOOST_CHECK_EQUAL(int_type{3}.binomial(-1),0);
-		BOOST_CHECK_EQUAL(int_type{10}.binomial(-1),0);
-		BOOST_CHECK_EQUAL(int_type{-3}.binomial(-3),1);
-		BOOST_CHECK_EQUAL(int_type{-1}.binomial(-1),1);
-	}
-};
-
-BOOST_AUTO_TEST_CASE(mp_integer_binomial_test)
-{
-	boost::mpl::for_each<size_types>(binomial_tester());
-	// Check the ints.
-	using int_type = integer;
-	BOOST_CHECK((has_binomial<int,int>::value));
-	BOOST_CHECK_EQUAL(math::binomial(4,2),math::binomial(int_type(4),2));
-	BOOST_CHECK((has_binomial<char,unsigned>::value));
-	BOOST_CHECK_EQUAL(math::binomial(char(4),2u),math::binomial(int_type(4),2));
-	BOOST_CHECK((has_binomial<long long,int>::value));
-	BOOST_CHECK_EQUAL(math::binomial(7ll,4),math::binomial(int_type(7),4));
-	BOOST_CHECK((std::is_same<decltype(math::binomial(7ll,4)),int_type>::value));
-	BOOST_CHECK_EQUAL(math::binomial(-7ll,4u),math::binomial(int_type(-7),4));
-	// Different bits sizes.
-	BOOST_CHECK((!has_binomial<mp_integer<16>,mp_integer<32>>::value));
-	BOOST_CHECK((!has_binomial<mp_integer<32>,mp_integer<16>>::value));
-}
-
-struct sin_cos_tester
-{
-	template <typename T>
-	void operator()(const T &)
-	{
-		typedef mp_integer<T::value> int_type;
-		BOOST_CHECK_EQUAL(math::sin(int_type()),0);
-		BOOST_CHECK_EQUAL(math::cos(int_type()),1);
-		BOOST_CHECK_THROW(math::sin(int_type(1)),std::invalid_argument);
-		BOOST_CHECK_THROW(math::cos(int_type(1)),std::invalid_argument);
-		BOOST_CHECK((std::is_same<int_type,decltype(math::cos(int_type{}))>::value));
-		BOOST_CHECK((std::is_same<int_type,decltype(math::sin(int_type{}))>::value));
-		BOOST_CHECK(has_sine<int_type>::value);
-		BOOST_CHECK(has_cosine<int_type>::value);
-	}
-};
-
-BOOST_AUTO_TEST_CASE(mp_integer_sin_cos_test)
-{
-	boost::mpl::for_each<size_types>(sin_cos_tester());
 }
 
 struct partial_tester
